Upversion python-openstacksdk to debian/1.0.1-2
As part of the STX-Openstack upversion to ANTELOPE, the python-
openstacksdk is being upversioned to the debian/1.0.1-2 branch.
openstacksdk v1.0.1 is the latest supported by Antelope [1].

The URL tag is being set to `debian/1.0.1-2` as it is the latest tag
for the 1.0.1 version on the Debian salsa mirror [2]

This package is being upversioned as using the old version (0.50.0-6)
was causing the python-openstackclient package to fail to build.

This change also adds the sha256sum field to the downloader path, as it
is being recommended to use it in the place of the md5sum.

[1] https://releases.openstack.org/antelope/#antelope-openstacksdk
[2] https://salsa.debian.org/openstack-team/libs/python-openstacksdk/-/tags

Test Plan:
PASS: Build the python-openstacksdk package

Story: 2010715
Task: 48137

Change-Id: I7a129d6552b60ef7d7c6f57b617c4a82149965a7
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>