Add master taint toleration in clients helm chart
It has been observed that, on Standard systems, `stx-openstack` is
failing to apply due to a non-tolerated taint in the stand-by
controller (node-role.kubernetes.io/master).

Although this is already being addressed in a separate revert change for
the majority of the helm charts [1], due to the fact that the `clients`
helm chart is relatively new, i.e., it wasn't included in the original
change that is being reverted, it requires a change of its own.

[1] https://review.opendev.org/c/starlingx/openstack-armada-app/+/886831

Test Plan (on Standard):
PASS - Build stx-openstack-helm-fluxcd package
PASS - Build stx-openstack helm charts
PASS - Upload/apply stx-openstack
PASS - Remove/delete stx-openstack

Partial-Bug: 2024909

Change-Id: I075180127b70ad47b111ba06a1e5d6b33e845330
Signed-off-by: Luan Nunes Utimura <LuanNunes.Utimura@windriver.com>