Upversion python-novaclient to 18.3.0-2
As part of the STX-Openstack upversion to ANTELOPE, the python-
novaclient is being upversioned to version 18.3.0-2, which is the
latest supported on ANTELOPE [1].

All patches were updated in order to make the package buildable in the
new version.

This change also adds the sha256sum field to the downloader path, as it
is being recommended to use it in the place of the md5sum.

[1] https://releases.openstack.org/antelope/#antelope-python-novaclient

Test Plan:
PASS: Build the python-novaclient package

Story: 2010715
Task: 48136

Change-Id: Idaa50c75a3681ca764eb7c644dcc999d7425e6c5
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>