Upversion python-neutronclient to 9.0.0-2
As part of the STX-Openstack upversion to ANTELOPE, the python-
neutronclient is being upversioned to version 9.0.0-2, which is the
latest supported on ANTELOPE [1].

All patches were updated in order to make the package buildable in the
new version.

This change also adds the sha256sum field to the downloader path, as it
is being recommended to use it in the place of the md5sum.

[1] https://releases.openstack.org/antelope/#antelope-python-neutronclient

Test Plan:
PASS: Build the python-neutronclient package

Story: 2010715
Task: 48135

Change-Id: I01a50f0e5b86dbe7acbbecf40dd3d095b2542f5d
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>