Upversion python-keystoneclient to 5.1.0-2
As part of the STX-Openstack upversion to ANTELOPE, the python-
keystoneclient is being upversioned to version 5.1.0-2, which is the
latest supported on ANTELOPE [1].

All patches were updated in order to make the package buildable in the
new version.

This change also adds the sha256sum field to the downloader path, as it
is being recommended to use it in the place of the md5sum.

[1] https://releases.openstack.org/antelope/#antelope-python-keystoneclient

Test Plan:
PASS: Build the python-keystoneclient package

Story: 2010715
Task: 48134

Change-Id: I4138d524cc63d383f046a24f4903d6a9d50b62ef
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>