Upversion python-ironicclient to 5.1.0-2
As part of the STX-Openstack upversion to ANTELOPE, the python-
ironicclient is being upversioned to version 5.1.0-2, which is the
latest supported on ANTELOPE [1].

All patches were updated in order to make the package buildable in the
new version.

This change also adds the sha256sum field to the downloader path, as it
is being recommended to use it in the place of the md5sum.

[1] https://releases.openstack.org/antelope/#antelope-python-ironicclient

Test Plan:
PASS: Build the python-ironicclient package

Story: 2010715
Task: 48133

Change-Id: I6b80f5725e6fb199b3909bd715a396ad93955b4d
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>