Upversion python-heatclient to 3.2.0-2
As part of the STX-Openstack upversion to ANTELOPE, the python-
heatclient is being upversioned to version 3.2.0-2, which is the
latest supported on ANTELOPE [1].

All patches were updated in order to make the package buildable in the
new version.

This change also adds the sha256sum field to the downloader path, as it
is being recommended to use it in the place of the md5sum.

[1] https://releases.openstack.org/antelope/#antelope-python-heatclient

Test Plan:
PASS: Build the python-heatclient package

Story: 2010715
Task: 48132

Change-Id: I974875e6fa37a54ad5a0398a0323ed414fea908a
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>