Upversion python-glanceclient to v4.3.0-2
As part of the STX-Openstack upversion to ANTELOPE, the python-
glanceclient is being upversioned to version 4.3.0-2, which is the
latest supported on ANTELOPE [1].

All patches were updated in order to make the package buildable in the
new version.

This change also adds the sha256sum field to the downloader path, as it
is being recommended to use it in the place of the md5sum.

[1] https://releases.openstack.org/antelope/#antelope-python-glanceclient

Test Plan:
PASS: Build the python-glanceclient package

Story: 2010715
Task: 48131

Change-Id: I72574a2ced8edfe495f236a14c0dedcfc4949297
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>