Upversion python-aodhclient to v3.2.0-2
As part of the STX-Openstack upversion to ANTELOPE, the python-
aodhclient is being upversioned to version 3.2.0-2, which is the
latest supported on ANTELOPE [1].

All patches were updated in order to make the package buildable in the
new version.

This change also adds the sha256sum field to the downloader path, as it
is being recommended to use it in the place of the md5sum.

[1] https://releases.openstack.org/antelope/#antelope-python-aodhclient

Test Plan:
PASS: Build the python-aodhclient package

Story: 2010715
Task: 48128

Change-Id: Ie0f475c551dc240810ee51ca1f5e72acc462e3e9
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>