Upversion python-barbicanclient to v5.5.0-2
As part of the STX-Openstack upversion to ANTELOPE, the python-
barbicanclient is being upversioned to version 5.5.0-2, which is the
latest supported on ANTELOPE [1].

All patches were updated in order to make the package buildable in the
new version.

This change also adds the sha256sum field to the downloader path, as it
is being recommended to use it in the place of the md5sum.

[1] https://releases.openstack.org/antelope/#antelope-python-barbicanclient

Test Plan:
PASS: Build the python-barbicanclient package

Story: 2010715
Task: 48129

Change-Id: I08d9d820ea4ab8d568fcd262a987bf8506c56840
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>