Upversion python-cinderclient to 9.3.0-1
As part of the STX-Openstack upversion to ANTELOPE, the python-
cinderclient is being upversioned to version 9.3.0-1, which is the
latest supported on ANTELOPE [1].

All patches were updated in order to make the package buildable in the
new version.

This change also adds the sha256sum field to the downloader path, as it
is being recommended to use it in the place of the md5sum.

Note that in the new version the /v2 directory is no longer used.

[1] https://releases.openstack.org/antelope/#antelope-python-cinderclient

Test Plan:
PASS: build-pkgs -c -p python-cinderclient

Story: 2010715
Task: 48130

Change-Id: I80e733e64f3c22222c52ac1068f841efaf9e9400
Signed-off-by: Romulo Leite <romulo.leite@windriver.com>