Upversion base openstack-helm-infra (6b6ca9e2)
This change upversion our base commit SHA for openstack-helm-infra,
in order to be more updated in comparison with
openstack/openstack-helm-infra main branch. It also ports all
StarlingX specific patches on top of it, dropping patches that are no
longer needed and updating what needs to be updated in order to be
applied on top of the new base SHA. Taking this chance to also update
the numbering/ordering of stx patches.
Chartmuseum build dependency was removed since a local helm chart server
is not really required for building openstack-helm-infra.

Helm Releases are updated with newest version of each helm chart.

Since the helm-toolkit chart was upversioned (0.2.19 -> 0.2.53), it is
also reflected on both the openstack-helm and stx-openstack-helm-fluxcd
build instructions.

Added ingress and error_pages images tags to both Ingress and MariaDB
static overrides in order to use the tags that are already being used by
the platform nginx-ingress-controller application.
ingress-nginx/controller was on v1.5.1 in the newest charts, was updated
to v1.7.0 that is already used. defaultbackend was on 1.4 in the newest
charts, was updated to -amd64:1.5.
Both are upgrades and not downgrades, so we should be fine and aligned
with images already delivered by default platform applications.

Finally, a new patch had to be developed (0016) in order to Add the
ServiceAccount creation back to openvswitch Daemonset definition.
Originally openvswitch chart had two daemonset definitions: for the
db and for the server, but recently both were merged into a single
daemonset [1] and the template inclusion was dropped during this merge.
This fix will also be proposed to openstack/openstack-helm-infra [2]

[1] https://github.com/openstack/openstack-helm-infra/commit/73e2b3322d3cc8ca4ee8453dd612266589d734b1
[2] https://review.opendev.org/c/openstack/openstack-helm-infra/+/888504

Test Plan:
PASS - Build openstack-helm-infra package
PASS - Build openstack-helm package
PASS - build-pkgs -c -a
PASS - Build stx-openstack application tarball
PASS - Apply/Remove stx-openstack
PASS - Apply/Remove stx-openstack with TLS enabled
PASS - Run stx-openstack sanity tests (AIO-SX)*
* No major issue found, 2 TCs failed but does not seem related to this
  upversion and if a consistent failure is confirmed a launchpad can be
  created later.

Story: 2010715
Task: 47887

Depends-On: https://review.opendev.org/c/starlingx/SDO-rv-service/+/886739

Signed-off-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>
Change-Id: Ie83075dbe3f60f43be7f3eb5c4870a2745a358ed