Upversion stx-glance image to Openstack Antelope
As part of the STX-Openstack upversion to Antelope, the stx-glance
image is being upversioned to the Antelope (stable/2023.1) branch [1]

Due to an issue with the wheels version (launchpad #2027616) [2], we're unable to build the images with the changes of this review, because of that, the test plan for this review will not contemplate the build of the image, and testing it on an environment with STX-Openstack. Only after this launchpad is worked on, we can proceed with the test plan, but, considering that the only possible option for the PROJECT_REF of this image is the `stable/2023.1` branch (in the context of the Antelope upversion), and that this is the only change this review is doing, it's ready for review.

It's worth noticing that these images are not being build in the official StarlingX build process, so there's no impact on the build itself.

[1] https://opendev.org/openstack/glance/src/branch/stable/2023.1
[2] https://bugs.launchpad.net/starlingx/+bug/2027616

Test Plan:
PASS: Make sure the branch specified in the is used when building the image PROJECT_REF

Story: 2010715
Task: 48121

Change-Id: Ia44e0caf24687d17c3477f1e5151d1457539429b
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>