Add UPGRADE_PIP_PACKAGES parameter to Images
This change adds UPGRADE_PIP_PACKAGES=pip argument to
stx-openstack images. This change will be necessary
when the images start to be build with Antelope version.

This parameter (was introduced in [1]) was added only
on the images that were facing issues on dependencies
versions resolution. The images that will receive this
parameter are:

- stx-aodh
- stx-cinder
- stx-horizon

On this change we also change the images file from this
repo PIP_PACKAGES disposal for allow better reading of
each file dependencies. Even images that are getting
the UPGRADE_PIP_PACKAGES parameter.

It is important to notice that this change alone won't
allow the images to be build, as we still need changes
in the build tools dependencies.

[1] https://review.opendev.org/c/starlingx/root/+/870429

Test Plan:
PASS: Manually change the build-tools repo with the
      necessary upversions of constraints and locally
      build the images.
PASS: Notice the PIP dependencies being resolved

Story: 2010715
Task: 48429

Signed-off-by: Romulo Leite <romulo.leite@windriver.com>
Change-Id: I2b1b4a728575bd4ec939ddfd7ac7a9538182b8b1