Removing debian_iso_image.inc file
There is no need of including stx-openstack-helm-fluxcd package to
StarlingX ISO, since this package is used only by the script that
generates the application tarball and this application is not delivered
along with the platform by default.

Having this package installed has no side-effects, but it installs
unnecessary files to /usr/lib on application, fluxcd and helm
directories.

TEST PLAN:
PASS - Successfully build a StarlingX ISO

Closes-Bug: 2028959

Signed-off-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>
Change-Id: I342266c0932bb7853aa5a377b3a5663677858b27