Enable TLS for the containerized clients
After the introduction of the containerized clients by [1], it was
noted that it was not possible to use them properly if the system had
HTTPS enabled.

This changes enabled the containerized clients to work when HTTPS is
enabled, without any intervention from the user, in the same way as the
platform clients work. For that, a secret had to be created
(secret-ingress-tls.yaml) containing the certificates information, and
this secret had to be mounted on the clients container. For this secret
to be created, a new job had to be created (job-boostrap.yaml) which
required some modifications to the values and static overrides for this
helm chart.

Besides that, some changes were required to the k8sapp-openstack, to
allow dynamic modification of the overrides for this chart when HTTPS
is enabled.

[1] https://review.opendev.org/c/starlingx/openstack-armada-app/+/885603

Test Plan (builds):
PASS: Build stx-openstack-helm-fluxcd and python3-k8sapp-openstack
PASS: Build Helm charts

Test Plan (HTTPS disabled):
PASS: Upload / apply the STX-Openstack application
PASS: Verify that all clients resources are up and running
PASS: Verify that none of the TLS resources were created (secret,
      volume, volume mount)
PASS: Verify that the containerized clients work as intended (issue
      an `openstack user list` after sourcing the /var/opt/openstack/
      admin-openrc file)
PASS: Remove / delete the STX-Openstack application

Test Plan (HTTPS enabled):
PASS: Upload / apply the STX-Openstack application
PASS: Verify that the required system overrides are applied
PASS: Verify that all clients resources are up and running
PASS: Verify that the TLS resources were created (secret, volume,
      volume mount)
PASS: Verify that the containerized clients work as intended (issue
      an `openstack user list` after sourcing the /var/opt/openstack/
      admin-openrc file)
PASS: Remove / delete the STX-Openstack application

Story: 2010774
Task: 48411

Change-Id: I8ff9170b55ae5399553a736ce857b50cc7866dc4
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>