clients: Make working dir. dyn. and improve chart
In the current state of the `clients` helm chart, the OpenStack clients'
working directory path is *static* and located at `/var/opt/openstack`.

While this directory is suitable for uploading small files -- such as
small Glance images -- it can quickly run out of space, as its maximum
capacity is, approximately, 20GB. This might not be enough for a single
user, and even worse, it wouldn't be for multiple users.

Therefore, the OpenStack clients' working directory needs to be
*dynamic*, which is the primary goal of this change.

With this change, the system administrator can change the default
clients' working directory by simply performing a helm override and
reapplying the app, such as:

  $ system helm-override-update stx-openstack clients openstack \
    --reuse-values --set workingDirectoryPath=/var/opt/another-directory

  $ system application-apply stx-openstack

During the re-apply, if the default working directory didn't contain
additional files in user subdirectories, it's removed to avoid leftovers
(in the same way as after removing the application).

For custom working directories, however, the handling is different.

They are *never* removed by the application, even during re-applies or
after application removal. The reason we don't remove it is because this
custom directory might be shared with other users and/or applications,
and, therefore, potentially have important files that must be preserved.

This change also improves the overall quality of the helm chart, by:
  * Fixing minor bugs;
  * Improving in-code documentation;
  * Raising exceptions on critical failures;
  * Removing functions and modules that are no longer needed;
  * Replacing None-returning functions with Boolean-returning functions,
    to better indicate success or failure status.

Test Plan:

PASS - Build python3-k8sapp-openstack package
PASS - Build stx-openstack-helm-fluxcd package
PASS - Build stx-openstack helm charts

Without exceptions:
-------------------
  PASS - Upload/apply stx-openstack
  PASS - Verify that the working directory was created at
         /var/opt/openstack
  PASS - Perform a helm override:
         $ system helm-override-update stx-openstack clients openstack \
           --reuse-values --set \
           workingDirectoryPath=/var/opt/another-directory
  PASS - Verify that the working directory was created at
        /var/opt/another-directory

With exceptions:
----------------
  PASS - Upload *faulty* stx-openstack
         This faulty version of the application raises, on purpose,
         a KubeAppApplyFailure exception during the creation of
         application specific resources
  PASS - Try to apply stx-openstack and verify that it immediately
         fails and goes to "apply-failed" state

PASS - Remove/delete stx-openstack

Story: 2010774
Task: 48410

Change-Id: Ieb06a93c669d362ecd30bff79947ee4c015cf2b4
Signed-off-by: Luan Nunes Utimura <LuanNunes.Utimura@windriver.com>