Remove RetryFilter
During the Upversion to Antelope it was noticed that
the nova pod was failing to start because of this
filter added during helm-override. The Retryfilter
was was deprecated in the Train release (20.0.0) and
REMOVED in Victoria (22.4.0)[1] and wont work on Antelope
and is not currently being used in Ussuri as well.

[1] https://docs.openstack.org/releasenotes/nova/victoria.html

Test plan:
PASS - build stx-nova image (Antelope)
PASS - upload custom image and push it to local registry
PASS - apply stx-openstack (with antelope images)
       and check pod health
PASS - helm-override nova chart (Ussuri)
PASS - apply stx-openstack (With ussuri images)
       and check pods health

Story: 2010715
Task: 48600

Signed-off-by: Romulo Leite <romulo.leite@windriver.com>
Change-Id: Iebe82d2fb9c90907ed6b45b6863870e675cbaa54