Update stx-openstackclients image definition
While testing the upversioned packages for this stx-openstackclients
image, it was noted that the build of it was failing due to some issues
with the dependencies for it, and also that some packages were missing
or unnecessary.

On the Victoria release for Openstack, the python-ceilometerclient was
removed from the project [1], and now that this image is going to be
build on the Antelope (stable/2023.1) release, this package is no longer
required for this image.

Also, it was noted that the python-aodhclient (called as simply
aodhclient) was not present on this image, probably because this image
was not maintained in the past as both the platform and the
STX-Openstack application shared the clients. Now that the clients for
the STX-Openstack application are containerized via this image, we have
to add this client on it.

Regarding the issue with this image's packages dependencies versions,
it was solved by adding the `UPGRADE_PIP_PACKAGES=pip` parameter.

[1] https://opendev.org/openstack/python-ceilometerclient

Test Plan:
PASS: Build all clients packages (aodh, cinder, glance, heat, keystone,
      neutron, nova, barbican and openstack clients)
PASS: Build stx-openstackclients image
PASS: Run image locally and verify that the aodh client is present, and
      the ceilometer client is not.
PASS: Apply the STX-Openstack app with Helm overrides pointing to this
      custom-build image
PASS: Make sure the clients are accessible by sourcing `/var/opt/openstack/admin-openrc` and issuing any openstack command

Story: 2010715
Task: 48137

Change-Id: I86fb4a48c68f3a76d3fdf4c82421fe9e95ccd5a8
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>