clients: Fix host-specific config overrides
It has been observed that, on Standard systems, users are unable to
interact with containerized clients due to the fact that the `clients`
pods are being initialized without the `controller-` suffix in their
names. These names are relevant for the clients wrapper script to be
able to find and execute commands inside the clients containers [1].

The reason pods are being initialized with wrong names on Standard
systems is because the host-specific overrides are being generated for
hosts with `WORKER` personality instead of `CONTROLLER`, which makes no
sense as the containerized clients were designed to *only* run on
controller nodes.

Therefore, the proposed fix is straightforward and only requires us to
change the host personality required to generate host-specific
overrides.

[1] https://opendev.org/starlingx/openstack-armada-app/src/commit/4c238a8063c27815954c3aa151cdcae7dd5f67cb/stx-openstack-helm-fluxcd/stx-openstack-helm-fluxcd/helm-charts/clients/templates/bin/_clients-wrapper.sh.tpl#L52-L55

Test Plan (on Standard):
PASS - Build python3-k8sapp-openstack package
PASS - Build stx-openstack-helm-fluxcd package
PASS - Build stx-openstack helm charts
PASS - Upload/apply stx-openstack
PASS - Source `/var/opt/openstack/admin-openrc` and check that the
       containerized clients are responding to commands as expected
PASS - Remove/delete stx-openstack

Closes-Bug: 2032382

Change-Id: I65b3fb64046567d18abb5335042a3df3b887d7c3
Signed-off-by: Luan Nunes Utimura <LuanNunes.Utimura@windriver.com>