Uncomment Openstack Debian packages and images
With the new STX-Openstack manifest [1], we can now uncomment the
packages in this repository so they can be build. These were previously
commented as they're also present on [2] and having duplicate packages
would cause the build to fail, or grab the incorrect packages.

This change also updates the build layer for this repo to the
openstack layer, that is being created by [3]

Note that since this change is on the `f/antelope` branch, it will not
impact the default manifest, nor the default StarlingX build.

[1] https://review.opendev.org/c/starlingx/manifest/+/891875
[2] https://opendev.org/starlingx/upstream/src/branch/master/debian_pkg_dirs
[3] https://review.opendev.org/c/starlingx/tools/+/892166

Test Plan:
PASS: Build all packages from this repository
PASS: Build all images from this repository

Depends-On: https://review.opendev.org/c/starlingx/manifest/+/893244
Depends-On: https://review.opendev.org/c/starlingx/manifest/+/893245

Story: 2010797
Task: 48691

Change-Id: I24677513f42f92258162ea09fe7e6c11b52093a1
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>