Add configurable control for Horizon sessions
Story/2010715 [1] is upversioning OpenStack services and packages to the
2023.1 (Antelope) release. Since the Horizon control for simultaneous
sessions of the same user [2] was introduced as part of OpenStack Yoga
release [3], stx-openstack application can now make use of such feature.

Users might be interested on altering this new Horizon configuration, so
the manifest static-override should have a key (to be overridden) for
such goal. This change adds to the "raw" configuration the
"simultaneous_sessions" key that can assume one of the two values:
  * simultaneous_session='allow': A user can login to several Horizon
  dashboards simultaneously (current behavior, default new config value)
  * simultaneous_session='disconnect': A user can NOT login to several
  Horizon dashboards simultaneously, only one active session (latest)
  is allowed. Any previous session will be disconnected after a
  subsequent successful login.

Default value for the stx-openstack will be "allow" for now.
This change implements now what was originally proposed in [4], but at
that time (i.e., in OpenStack Ussuri) there was no reason for adding a
configurable control to a feature that was not yet available in
stx-openstack.

[1] https://storyboard.openstack.org/#!/story/2010715
[2] https://blueprints.launchpad.net/horizon/+spec/handle-multiple-login-sessions-from-same-user-in-horizon
[3] https://docs.openstack.org/releasenotes/horizon/yoga.html
[4] https://review.opendev.org/c/starlingx/openstack-armada-app/+/814491

TEST PLAN:
PASS - Build stx-openstack-helm-fluxcd package
PASS - Build stx-openstack helm charts
PASS - Upload/apply stx-openstack (AIO-SX virtual)
PASS - Verify horizon config from inside the pod
PASS - Apply a helm-override to change the config value to "disconnect"
     conf:
       horizon:
         local_settings:
           config:
             simultaneous_sessions: 'disconnect'
PASS - Verify horizon config from inside the pod
PASS - Remove/delete stx-openstack

*NOTE: Initial tests indicate that might be a bug on this feature when
using the Antelope stx-horizon image. But that would be a Horizon code
and not an issue with this configuration mechanism, that was prove to be
working fine. Feature issues will be addressed by separate launchpads.

Story: 2009304
Task: 43668

Signed-off-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>
Change-Id: I74a00777e000b20acb0c1e030d947d907b90a4f9