Update Horizon SSL Cert path
It was observed that, when using the STX-Openstack app with HTTPS
enabled, it was not possible to visualize / edit / create images using
Horizon. This was caused because the SSL certificate (ca.crt) volume
mount path defined in Horizon's configuration [1] was not the same as
the default one created when using the Helm toolkit [2].

As specified in [3], one possible solution was to define a path to the
TLS secret to be created, but that caused other services in Horizon to
fail, so instead of doing that, the solution was to change Horizon's
configuration to search the ca.crt file in the correct path.

[1] https://opendev.org/openstack/openstack-helm/src/branch/master/horizon/values.yaml#L469
[2] https://opendev.org/openstack/openstack-helm-infra/src/branch/master/helm-toolkit/templates/snippets/_tls_volume_mount.tpl#L56
[3] https://opendev.org/openstack/openstack-helm-infra/src/branch/master/helm-toolkit/templates/snippets/_tls_volume_mount.tpl#L30

Test Plan:
PASS: Build openstack-helm and stx-openstack-helm-fluxcd
PASS: Build STX-Openstack tarball
PASS: Upload / apply STX-Openstack with HTTPS enabled
PASS: Create an image by using the Glance CLI
PASS: On Horizon, under Project -> Compute -> Images, verify that the
      image is present
PASS: Create an image via Horizon

Closes-Bug: 2032980

Change-Id: If34c0d4a287804e4fce0e5f70bb97feb330be174
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>