Add `get_namespace` function to OpenstackBaseHelm
After the introdution of [1], it was noted that Helm overrides were not
working properly for STX-Openstack. Althought they show up on
helm-override-show, they we're not being passed to Helm and Kubernetes.
This was caused because the [1] change updated the way that the
overrides are applied, requiring the `get_namespace` function on the
plugin to define where to apply these overrides, and, if this function
did not exist, it would default to the `default` namespace.

This change fixes that by adding the `get_namespace` function to the
OpenstackBaseHelm plugin, which is extended by the other STX-O plugins.

This change also does the following:

1) Updates the base class for ingress, memcached and garbd from
   BaseHelm to OpenstackBaseHelm, as openstack.BaseHelm is the old base
   class for the armada app and probably should be cleaned now that we
   no longer support Armada.
2) Remove ingress customizations for the kube-system namespace, as
   they're not used anymore and can be removed without any impact.

[1] https://review.opendev.org/c/starlingx/config/+/887430

Test Plan:
PASS: Build python3-k8sapp-openstack and stx-openstack-helm-fluxcd
PASS: Build STX-Openstack tarball
PASS: Upload / apply STX-Openstack on an AIO-SX machine with HTTPS
      enabled
PASS: Verify that there's only one ingress pod per component
PASS: After passing helm overrides, verify that they're applied
      (Example: change number of pods from 1 to 2)
PEND: Confirm overrides work for every Openstack component

Closes-Bug: 2036633

Change-Id: Id01a7367335213015531e1e9186b2b1cc677ef8e
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>